refactor(fsd): use cart public api and drop deep imports
