+++ src/features/cart/index.js
@@ -1,0 +1,2 @@
+export { default as useCartStore } from './model/cartStore';
+export { selectCartCount, selectCartItems, selectCartTotal } from './model/selectors';
